tests/requirements-test.txt
# Testing Framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-xdist==3.3.1
pytest-env==1.1.1
pytest-randomly==3.15.0

# Test Utils
factory-boy==3.3.0
Faker==20.1.0
coverage==7.3.2
freezegun==1.2.2
requests-mock==1.11.0
aioresponses==0.7.4
asynctest==0.13.0
mock==5.1.0
responses==0.24.1

# Development Dependencies
black==23.7.0
flake8==6.1.0
mypy==1.5.1
isort==5.12.0 